mainへ40Hアラートの追加（テスト用設定値）
